Ability timeToReady() overeride
Allows overriding of timeToReady() in cases where we need to manually intervene. Example can be seen in Devastation Evoker's Pyre ability, where this override fixes the rotation and the UI components. Issue is caused by Blizzard registering the ability as on cooldown in relation to resource regeneration (such as essence), even though the addon is fully aware that resource will become available (due to a guaranteed proc) on the next spellcast.