fix: always show the correct texture for an action
If an ability can have different textures but is invoked in a priority
under only one name, then the display would show only the previous
texture instead if the action was the same as before.

Fix this behavior by also saving the previous texture shown in the
display and checking if it differs from the current texture to be
displayed.