fix: make the APL actions "maul" and "raze" work correctly
For Guardian druids in TWW2, Raze now directly replaces Maul if talented
into Raze, and Ravage replaces both if the Druid of the Claw Ravage buff
is up. For the APLs, we need to support `maul` and `raze` as separate
actions based on if Raze is talented, and the actions should show the
Ravage texture if the Ravage buff is up.

Remove "ravage" is a usable action since it's not supported by SimC and
shouldn't be used in an APL -- use `maul` or `ravage` instead and check
for `buff.ravage.up` if needed.

Implement a hidden `ravage_maul` action that encapsulates the
information for the Ravage ability and also has a handler function for
the secondary effects triggered by casting Ravage. This matches how SimC
implements this ability internally as well.