fix: simplify Ironfur modeling for guardian druid
Make the `ironfur_damage_threshold` state expression adjust for whether
the player is solo or in a group to centralize the computation in one
place.

Remove the `usable` check for `ironfur` that had extra checks for
whether it was okay to cast Ironfur. Those checks are already in the
default APL and are redundant.