fix: synchronize Metamorphosis code with Vengeance model
Copy code from Vengeance model to handle Metamorphosis with Demonsurge.
This introduces `demonsurge_hardcast` to Havoc and tracks it properly to
check for when various spells should be empowered.

Be more accurate with adding stacks of Demonsurge only when talented
into Demonic Intensity. This mirrors how the buff is handled in the
Vengeance model.

Fix the shared cooldowns for `eye_beam` and `abyssal_gaze`.

Fix the amount of haste granted by Metamorphosis.