Shadow Priest - Fix Void Torrent interrupt_if condition
I think the `!talent.void_torrent` was struggling to go through the interpreter into the `if modifiers.interrupt_if and modifiers.interrupt_if() then`.

Changing it to our normal internal syntax fixes it, and it recommends the channel break as soon as your GCD is ready.

Fixes https://github.com/Hekili/hekili/issues/5046